chore: general Storybook improvements
