+++ .storybook/main.ts
@@ -1,12 +1,13 @@
 import type { StorybookConfig } from '@storybook/react-webpack5';
 const config: StorybookConfig = {
 	stories: [
-		'../stories/**/*.mdx',
-		'../stories/**/*.stories.@(js|jsx|ts|tsx)',
+		'../stories/pages/**/*.mdx',
+		'../stories/docs/**/*.stories.@(js|jsx|ts|tsx)',
 	],
 	addons: [
 		'@storybook/addon-links',
 		'@storybook/addon-essentials',
+		'@storybook/addon-docs',
 		'@storybook/addon-interactions',
 		'@storybook/addon-a11y',
 	],
