+++ .storybook/manager.js
@@ -1,0 +1,6 @@
+import { addons } from '@storybook/addons';
+import { themes } from '@storybook/theming';
+
+addons.setConfig({
+	theme: themes.dark,
+});
