+++ package.json
@@ -32,7 +32,9 @@
 		"@babel/preset-react": "^7.22.3",
 		"@babel/preset-typescript": "^7.21.5",
 		"@storybook/addon-a11y": "^7.1.1",
+		"@storybook/addon-docs": "^7.2.0",
 		"@storybook/addon-essentials": "^7.0.23",
+		"@storybook/theming": "^7.2.0",
 		"@storybook/addon-interactions": "^7.0.23",
 		"@storybook/addon-links": "^7.0.23",
 		"@storybook/blocks": "^7.0.23",
@@ -59,6 +61,7 @@
 		"react-dom": "^18.2.0",
 		"react-icons": "^4.9.0",
 		"storybook": "^7.0.23",
+		"storybook-dark-mode": "^3.0.1",
 		"styled-components": "6.0.0-rc.3",
 		"ts-loader": "^9.4.3",
 		"typescript": "^5.0.2"
@@ -67,5 +70,8 @@
 		">0.2%",
 		"not dead",
 		"not op_mini all"
-	]
+	],
+	"dependencies": {
+		"polished": "^4.2.2"
+	}
 }
