+++ stories/pages/tokens/colors.stories.mdx
@@ -1,0 +1,10 @@
+import { Meta } from '@storybook/blocks';
+import { ColorsGrid } from '../components/colorGrid/index.tsx';
+
+<Meta title="Tokens/Colors" />
+
+# Colors
+
+Essas são as cores utilizadas na WDS Core
+
+<ColorsGrid />
