+++ stories/pages/tokens/spacing.stories.mdx
@@ -1,0 +1,10 @@
+import { Meta } from '@storybook/blocks';
+import { TokenTable } from '../components/tokenTable';
+
+<Meta title="Tokens/Spacing" />
+
+# Spacing
+
+Esses são os valores de espaçamento usados na WDS Core
+
+<TokenTable />
